pyproject.toml
[project]
name = "projet-2-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.5",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
